refactor: Use InfluxQL implementation for simplifying time ranges
