+++ iox_query_influxql/src/plan/mod.rs
@@ -2,9 +2,10 @@ mod error;
 mod expr_type_evaluator;
 mod field;
 mod field_mapper;
-mod influxql_time_range_expression;
 mod ir;
 mod planner;
+mod planner_rewrite_expression;
+mod planner_time_range_expression;
 mod rewriter;
 mod test_utils;
 mod util;
